ML_API/requirements.txt
# API
fastapi
uvicorn
python-multipart

# ML (БЕЗ torch!)
transformers
openai-whisper
